feat: adding initial structure table
