+++ front-busadmin/src/components/Table/index.tsx
@@ -1,0 +1,5 @@
+export const Table = () => {
+  return (
+    <table></table>
+  );
+};
