+++ front-busadmin/src/utils/imagesPath.ts
